docs: fix tag ordering for pages
