+++ .eleventy.cjs
@@ -7,6 +7,7 @@ const pfeAssetsPlugin = require('@patternfly/pfe-tools/11ty/plugins/pfe-assets.c
 
 const customElementsManifestPlugin = require('@patternfly/pfe-tools/11ty/plugins/custom-elements-manifest.cjs');
 const orderTagsPlugin = require('@patternfly/pfe-tools/11ty/plugins/order-tags.cjs');
+const alphabetizeTagsPlugin = require('./docs/_plugins/alphabetize-tags.cjs');
 const todosPlugin = require('@patternfly/pfe-tools/11ty/plugins/todos.cjs');
 
 const markdownIt = require('markdown-it');
@@ -146,68 +147,17 @@ module.exports = function(eleventyConfig) {
     warningFileSize: 400 * 1000,
   });
 
-
   /**
    * Collections to organize by 'order' value in front matter, then alphabetical by title;
    * instead of by date
    */
-  // Tags to sort
-  const tagsToAlphabetize = [
-    // 'component',
-    'foundations',
-  ];
-
-  // Iterate over tags to sort
-  for (const tag of tagsToAlphabetize) {
-    eleventyConfig.addCollection(tag, function(collection) {
-      const currentCollection = collection.getFilteredByTag(tag);
-      // weights will have a key for each specified weight (the value of order),
-      // and an array of each item with that weight
-      const weights = {};
-      // Final sorted array of collection items
-      let sorted = [];
-      for (let index = 0; index < currentCollection.length; index++) {
-        const item = currentCollection[index];
-        // If order is set, remove it from currentCollection and add it to weights
-        if (item.data.order) {
-          const { order } = item.data;
-          if (!weights[order]) {
-            weights[order] = [];
-          }
-          const removedItem = currentCollection.splice(index, 1);
-          weights[order].push(removedItem[0]);
-        }
-      }
-
-      // Default non specified pages to a weigh tof 0 by
-      // adding what's remaining in currentCollection to weight 0
-      if (weights[0]) {
-        weights[0].concat(currentCollection);
-      } else {
-        weights[0] = currentCollection;
-      }
-
-      // Iterate over weights with multiple items and sort by title alphabetically
-      // @note The .sort() may need a sort handler that uses parseInt, but seems to be working?
-      const weightKeys = Object.keys(weights).sort();
-      for (const currentWeight of weightKeys) {
-        // Sort by title alphabetically
-        weights[currentWeight].sort(function(a, b) {
-          if (a.data.title < b.data.title) {
-            return -1;
-          }
-          if (a.data.title > b.data.title) {
-            return 1;
-          }
-          return 0;
-        });
-        // Append result to sorted array
-        sorted = sorted.concat(weights[currentWeight]);
-      }
-
-      return sorted;
-    });
-  }
+  eleventyConfig.addPlugin(alphabetizeTagsPlugin, {
+    tagsToAlphabetize: [
+      // 'component',
+      'foundations',
+      'getstarted',
+    ]
+  });
 
   eleventyConfig.setLibrary('md', markdownLib);
